refactor: prooject structure
