+++ src/components/GenerateEmbedCodeDialog.vue
@@ -75,7 +75,7 @@
 <script lang="ts" setup>
 import { onMounted, reactive, watch } from "vue";
 import toast from "../helpers/toast";
-import utils from "../helpers/utils";
+import utils from "../../common/utils";
 import useAppStore from "../store";
 import Dialog from "./Dialog.vue";
