+++ tsconfig.json
@@ -15,5 +15,5 @@
     "checkJs": false,
     "lib": ["esnext", "dom"]
   },
-  "include": ["src/**/*.ts", "src/**/*.vue", "scripts/*.ts", "packages/**/*.ts"]
+  "include": ["src/**/*.ts", "src/**/*.vue", "scripts/*.ts", "packages/**/*.ts", "common/**/*.ts", "types/*.ts"]
 }
