+++ types/chart.ts
@@ -1,0 +1,9 @@
+export type ChartMode = "Date" | "Timeline";
+
+export interface RepoStarData {
+  repo: string;
+  starRecords: {
+    date: string;
+    count: number;
+  }[];
+}
